carrito terminado

+++ Koala/Controllers/CartController.cs
@@ -32,13 +32,17 @@ namespace Koala.Controllers
                 var producto = productos.First(p => p.Id_Producto == item.IdProducto);
                 if (item != null)
                 {
+                    double precio = (double)(item.Cantidad * producto.Precio);
+                    double descuento = (100 - producto.Descuento) / 100;
+                    double precioConDescuento = precio * descuento;
                     lista.Add(new CartViewModel
                     {
+                        IdCarrto = item.IdCarrito,
                         Cantidad = item.Cantidad,
                         IdProducto = item.IdProducto,
                         NombreProducto = producto.Nombre,
                         Descuento = producto.Descuento,
-                        Precio = item.Cantidad * producto.Precio
+                        Precio = precioConDescuento
                     });
                 }
             }
@@ -67,8 +71,7 @@ namespace Koala.Controllers
             var productoEnCarrito = await _db.Carrito.Where(c => c.IdProducto == id).FirstOrDefaultAsync();
             if (productoEnCarrito != null)
             {
-
-                productoEnCarrito.Cantidad++; 
+                productoEnCarrito.Cantidad++;
                 _db.Entry(productoEnCarrito).State = EntityState.Modified;
             }
             else
@@ -85,12 +88,61 @@ namespace Koala.Controllers
             return RedirectToAction("Index");
         }
 
+        [HttpPost]
         public async Task<ActionResult> Delete(int? id)
         {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+            }
             Carrito carrito = await _db.Carrito.FindAsync(id);
+            if (carrito == null)
+            {
+                return HttpNotFound();
+            }
             _db.Carrito.Remove(carrito);
             await _db.SaveChangesAsync();
             return RedirectToAction("Index");
+        }
+
+        [HttpPost]
+        public async Task<ActionResult> MakeOrder()
+        {
+            var usuario = await GetUser();
+            var cliente = await _db.Clientes.Where(c => c.DNI_Cliente == usuario.DNI).FirstOrDefaultAsync();
+            var productos = await _db.Productos.ToListAsync();
+            if (cliente != null)
+            {
+                var articulosCarrito = await _db.Carrito.Where(c => c.IdCliente == cliente.Id_Cliente).ToListAsync();
+
+                var pedido = new Pedidos
+                {
+                    Cliente = cliente.Id_Cliente,
+                    Fecha_Pedido = DateTime.Now,
+                    Confirmado = DateTime.Now
+                };
+                _db.Pedidos.Add(pedido);
+                await _db.SaveChangesAsync();
+
+                foreach (var item in articulosCarrito)
+                {
+                    var producto = productos.Find(p => p.Id_Producto == item.IdProducto);
+                    _db.Línea_Pedido.Add(new Línea_Pedido
+                    {
+                        Pedido = pedido.Id_Pedido,
+                        Cantidad = item.Cantidad,
+                        Precio = producto.Precio,
+                        Nombre = producto.Nombre,
+                        Producto = producto.Id_Producto
+                    });
+                }
+                await _db.SaveChangesAsync();
+
+                _db.Carrito.RemoveRange(articulosCarrito);
+                await _db.SaveChangesAsync();
+                return RedirectToAction("Index", "Home");
+            }
+            return View("Index");
         }
     }
 }
